% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE25902.R
\docType{data}
\name{GSE25902}
\alias{GSE25902}
\title{GSE25902}
\format{A Summarized Experiment object with 19014 genes and 120 samples (24 cases and 96 controls).
The column outcome in the colData corresponds to the outcome that was used in the paper.}
\usage{
data(GSE25902)
}
\description{
This is a preprocessed hallmark data set with ALLOGRAFT_REJECTION as target pathway.
 A Genome U133 Plus 2.0 Array is utilized to analyze post transplant rejection in kidney tissue. The study was performed in an unpaired design.
}
\references{
Naesens, M., Khatri, P., Li, L., Sigdel, T. K., Vitalone, M. J., Chen, R., Butte, A. J., Salvatierra, O., and Sarwal, M. M. (2011). Progressive histological damage in renal allografts is associated with expression of innate and adaptive immunity genes. Kidney Int, 80, 1364–1376.
}
\keyword{datasets}
