% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE27034.R
\docType{data}
\name{GSE27034}
\alias{GSE27034}
\title{GSE27034}
\format{A Summarized Experiment object with 19014 genes and 36 samples (19 cases and 17 controls).
The column outcome in the colData corresponds to the outcome that was used in the paper.}
\usage{
data(GSE27034)
}
\description{
This is a preprocessed hallmark data set with COAGULATION as target pathway.
 A Genome U133 Plus 2.0 Array is utilized to analyze peripheral artery disease in blood tissue. The study was performed in an unpaired design.
}
\references{
Masud, R., Shameer, K., Dhar, A., Ding, K., and Kullo, I. J. (2012). Gene expression profiling of peripheral blood mononuclear cells in the setting of peripheral arterial disease. J Clin Bioinf , 2, 6.
}
\keyword{datasets}
