% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE27262.R
\docType{data}
\name{GSE27262}
\alias{GSE27262}
\title{GSE27262}
\format{A Summarized Experiment object with 19014 genes and 42 samples (20 cases and 22 controls).
The column outcome in the colData corresponds to the outcome that was used in the paper.}
\usage{
data(GSE27262)
}
\description{
This is a preprocessed hallmark data set with GLYCOLYSIS as target pathway.
 A Genome U133 Plus 2.0 Array is utilized to analyze non-small cell lung cancer in lung tissue. The study was performed in a paired design.
}
\references{
Wei, T.-Y. W., Juan, C.-C., Hisa, J.-Y., Su, L.-J., Lee, Y.-C. G., Chou, H.-Y., Chen, J.-M. M., Wu, Y.-C., Chiu, S.-C., Hsu, C.-P., et al. (2012). Protein arginine methyltransferase 5 is a potential oncoprotein that upregulates G 1 cyclins/cyclin-dependent kinases and the phosphoinositide 3-kinase/AKT signaling cascade.
}
\keyword{datasets}
