% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE33335.R
\docType{data}
\name{GSE33335}
\alias{GSE33335}
\title{GSE33335}
\format{A Summarized Experiment object with 17025 genes and 50 samples (25 cases and 25 controls).
The column outcome in the colData corresponds to the outcome that was used in the paper.}
\usage{
data(GSE33335)
}
\description{
This is a preprocessed hallmark data set with GLYCOLYSIS as target pathway.
 A Exon 1.0 ST Array is utilized to analyze gastric cancer in gastric tissue. The study was performed in a paired design.
}
\references{
Cheng, L., Wang, P., Yang, S., Yang, Y., Zhang, Q., Zhang, W., Xiao, H., Gao, H., and Zhang, Q. (2012). Identification of genes with a correlation between copy number and expression in gastric cancer. BMC Med Genomics, 5, 14.
}
\keyword{datasets}
