% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE40586.R
\docType{data}
\name{GSE40586}
\alias{GSE40586}
\title{GSE40586}
\format{A Summarized Experiment object with 18593 genes and 39 samples (21 cases and 18 controls).
The column outcome in the colData corresponds to the outcome that was used in the paper.}
\usage{
data(GSE40586)
}
\description{
This is a preprocessed hallmark data set with COMPLEMENT as target pathway.
 A Gene 1.0 ST Array is utilized to analyze bacterial meningitis in blood tissue. The study was performed in a unpaired design.
}
\references{
Lill, M., Taba, P., Kõks, S., Soomets, U., Lutsar, I., Schalkwyk, L. C., and Fernandes, C. (2013). Peripheral blood RNA gene expression profiling in patients with bacterial meningitis. Front Neurosci, 7, 33.
}
\keyword{datasets}
