% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE50635.R
\docType{data}
\name{GSE50635}
\alias{GSE50635}
\title{GSE50635}
\format{A Summarized Experiment object with 18605 genes and 49 samples (33 cases and 16 controls).
The column outcome in the colData corresponds to the outcome that was used in the paper.}
\usage{
data(GSE50635)
}
\description{
This is a preprocessed hallmark data set with INTERFERON_ALPHA_RESPONSE as target pathway.
 A Gene 1.0 ST Array is utilized to analyze systemic lupus erythematosus in blood tissue. The study was performed in a unpaired design.
}
\references{
Ko, K., Koldobskaya, Y., Rosenzweig, E., and Niewold, T. B. (2013). Activation of the interferon pathway is dependent upon autoantibodies in African-American SLE patients, but not in European-American SLE patients. Front Immunol, 4, 309.
}
\keyword{datasets}
