% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE8671.R
\docType{data}
\name{GSE8671}
\alias{GSE8671}
\title{GSE8671}
\format{A Summarized Experiment object with 12482 genes and 64 samples (32 cases and 32 controls).
The column outcome in the colData corresponds to the outcome that was used in the paper.}
\usage{
data(GSE8671)
}
\description{
This is a preprocessed hallmark data set with WNT_BETA_CATENIN_SIGNALING as target pathway.
 A Genome U133 Plus 2.0 Array is utilized to analyze colon cancer in colon tissue. The study was performed in a paired design.
}
\references{
Sabates-Bellver, J., Van der Flier, L. G., de Palo, M., Cattaneo, E., Maake, C., Rehrauer, H., Laczko, E., Kurowski, M. A., Bujnicki, J. M., Menigatti, M., et al. (2007). Transcriptome profile of human colorectal adenomas. Mol Cancer Res, 5, 1263–1275.
}
\keyword{datasets}
